533 move applicant optionsets into their own file
